Cleaned up directories, Refactored code, Removed deadclass Remove unused code Removed dockerfile
