+++ resource-service/src/main/java/com/project/resourceservice/ResourceApplication.java
@@ -1,13 +1,13 @@
-package com.project.mp3resourceservice;
+package com.project.resourceservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Mp3ResourceApplication {
+public class ResourceApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(Mp3ResourceApplication.class, args);
+        SpringApplication.run(ResourceApplication.class, args);
     }
 
 }
