+++ resource-service/src/main/java/com/project/resourceservice/client/RestClient.java
@@ -1,7 +1,7 @@
-package com.project.mp3resourceservice.client;
+package com.project.resourceservice.client;
 
 
-import com.project.mp3resourceservice.dto.Mp3MetadataDto;
+import com.project.resourceservice.dto.Mp3MetadataDto;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
