+++ resource-service/src/main/java/com/project/resourceservice/config/SongServiceProperties.java
@@ -1,4 +1,4 @@
-package com.project.mp3resourceservice.config;
+package com.project.resourceservice.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
