+++ resource-service/src/main/java/com/project/resourceservice/dto/Mp3FileResponseDto.java
@@ -1,4 +1,4 @@
-package com.project.mp3resourceservice.dto;
+package com.project.resourceservice.dto;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import lombok.Data;
