+++ resource-service/src/main/java/com/project/resourceservice/exception/CSVStringException.java
@@ -1,4 +1,4 @@
-package com.project.mp3resourceservice.exception;
+package com.project.resourceservice.exception;
 
 public class CSVStringException extends RuntimeException {
     public CSVStringException(String message) {
