+++ resource-service/src/main/java/com/project/resourceservice/repo/Mp3FileRepository.java
@@ -1,6 +1,6 @@
-package com.project.mp3resourceservice.repo;
+package com.project.resourceservice.repo;
 
-import com.project.mp3resourceservice.entity.Mp3File;
+import com.project.resourceservice.entity.Mp3File;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface Mp3FileRepository extends JpaRepository<Mp3File, Long> {
