+++ resource-service/src/main/java/com/project/resourceservice/service/Mp3FileService.java
@@ -1,18 +1,18 @@
-package com.project.mp3resourceservice.service;
+package com.project.resourceservice.service;
 
-import com.project.mp3resourceservice.client.RestClient;
-import com.project.mp3resourceservice.config.SongServiceProperties;
-import com.project.mp3resourceservice.dto.Mp3FileResponseDto;
-import com.project.mp3resourceservice.dto.Mp3IdListResponseDto;
-import com.project.mp3resourceservice.dto.Mp3MetadataDto;
-import com.project.mp3resourceservice.entity.Mp3File;
-import com.project.mp3resourceservice.exception.CSVStringException;
-import com.project.mp3resourceservice.exception.InvalidFileFormatException;
-import com.project.mp3resourceservice.exception.ResourceNotFoundException;
-import com.project.mp3resourceservice.exception.RestClientException;
-import com.project.mp3resourceservice.mapper.Mp3FileToResponseDtoMapper;
-import com.project.mp3resourceservice.repo.Mp3FileRepository;
-import com.project.mp3resourceservice.util.DurationFormatter;
+import com.project.resourceservice.client.RestClient;
+import com.project.resourceservice.config.SongServiceProperties;
+import com.project.resourceservice.dto.Mp3FileResponseDto;
+import com.project.resourceservice.dto.Mp3IdListResponseDto;
+import com.project.resourceservice.dto.Mp3MetadataDto;
+import com.project.resourceservice.entity.Mp3File;
+import com.project.resourceservice.exception.CSVStringException;
+import com.project.resourceservice.exception.InvalidFileFormatException;
+import com.project.resourceservice.exception.ResourceNotFoundException;
+import com.project.resourceservice.exception.RestClientException;
+import com.project.resourceservice.mapper.Mp3FileToResponseDtoMapper;
+import com.project.resourceservice.repo.Mp3FileRepository;
+import com.project.resourceservice.util.DurationFormatter;
 import org.apache.tika.exception.TikaException;
 import org.apache.tika.metadata.Metadata;
 import org.apache.tika.parser.ParseContext;
