+++ resource-service/src/main/java/com/project/resourceservice/util/DurationFormatter.java
@@ -1,4 +1,4 @@
-package com.project.mp3resourceservice.util;
+package com.project.resourceservice.util;
 
 import java.time.Duration;
